## Description
Added Mentor selection on the Matching Page of underdog, each user gets the mentors assigned based on the assignment and the mentor ID in the assignment object.

Fixes # (issue)

cleaned allot of code and refactored some, the code base had allot of issues in it, props not being passed correctly and using dummy data on the wrong page.

## Loom Video

https://www.loom.com/share/e9096227a5314e0a93ad9988e7a4274d

## Type of change

Please delete options that are not relevant.

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)

## Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have removed unnecessary comments/console logs from my code
- [x] I have made corresponding changes to the documentation if necessary (optional)
- [x] My changes generate no new warnings
- [x] I have checked my code and corrected any misspellings
- [x] No duplicate code left within changed files
- [x] Size of pull request kept to a minimum
- [x] Pull request description clearly describes changes made & motivations for said changes